requirements.txt
asgiref==3.5.2
attrs==22.1.0
Django==4.1.3
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
exceptiongroup==1.0.4
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.5
pyparsing==3.0.9
pytest==7.2.0
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
